Color improvements.
Use symbolic colors to define the background color of the window and for the
focus line. The main problem is that I can't use bg[NORMAL] in the spinbutton
to mean the background of the window, as I want to use it for the up/down
buttons.
Using symbolic colors for the focus_line also simplifies the checkbutton,
which should follow normal foreground colors, but always has a white focus
line.