sugar-artwork: stricter compiler flags
WARN_CFLAGS was already used in gtk/engine/Makefile.am, but not
initialized by configure.

For backwards compatibility, GCC's by default is very permissive.
In my case, it was silently miscompiling code which was calling
a non-existent GTK function. For this particular case, the
proposed WARN_CFLAGS should make the build fail early in this case.

I'm not proposing -Werror because it frequetly results in broken
builds for trivial reasons when distros update the toolchain.

Signed-off-by: Bernie Innocenti <bernie@codewiz.org>
Reviewed-by: James Cameron <quozl@laptop.org>
Tested-by: James Cameron <quozl@laptop.org>