Theme the LinkInfo label that is displayed in Browse at the bottom left
Set the color of the label and the background of it's parent the
overlay. Labels don't create their own window to render themselves
into. Instead, they draw themselves directly onto their parents
window, so we have to set the background-color of the parent in order
to set the labels' background color [1].

The patch does adjust the Browse entry names to be consistent.

[1] http://developer.gnome.org/gtk-faq/stable/x810.html

Signed-off-by: Simon Schampijer <simon@schampijer.de>