Fix artifacts in share button, SL #3387
When the SugarRadioToolButton is inside a SugarPaletteWindowWidget
and the state is active we make the background color button_grey. When a
GtkToolButton is inside of a SugarPaletteWindowWidget and the state is active
we make the background color transparent, otherwise this fills up all the space
around the SugarRadioToolButton.

For all the other appearences of a GtkToolButton we make the background color
button_grey when the state is active which fixes partly SL #3590.

Initial-patch: Carlos Garnacho <carlosg@gnome.org>
Signed-off-by: Simon Schampijer <simon@laptop.org>
Tested-by: Gonzalo Odiard <gonzalo@laptop.org>