Make the main toolbar the right size - SL #3388
Set to zero the GtkWidget-focus-padding which was changing the size of
the buttons, and was noticeable in the toolbar because the toolbar
buttons were changing the toolbar height.  Use the toolbutton_padding
variable instead.  It was used before only for the tool buttons
border-radius, but a look at the gtk2 theme gtkrc shows that its used
there for the inner-border, which is the padding in gtk3 CSS.  It
gained 1px to make the toolbar exactly the same height than in the
gtk2 theme.  I tested this with screenshots in both sugar-100 and
sugar-72.  I'm attaching them to the ticket.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>