Remove colors, borders and paddings set for all the widgets, fix Button borders - SL #3837 #3837
The * {} and *:active {} CSS rules apply to all the widgets so they
have to be overriden many times.  We should do that for each widget
when needed instead.

Also fix the border for the GtkButton.  We have a white outline in its
focused state.  This is implemented as a CSS border in the GTK3+
version.  But if the button doesn't have a border in the normal state,
we get a resize in the parent widget.  This patch adds a border that
is the same color than the button background for the normal state.
Also adds a white background to the active (pressed) state.

Also:

- fix the label in the Browse notebook, SL #3837
- use .menuitem class instead of GtkMenuItem

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>

----

v2: fix regressions in the toolbar buttons
v3: use .menuitem class, fix color for .button:active:focused