Set correct border-color for entries - SL #3389
Now the entries in the toolbar have a border color that is the same as
the background color, looking like in GTK2.  Entries with a
progressbar inside, like Browse activity URL entry, now have the same
style than in GTK2, with an outline.  See images attached to the
ticket for reference.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>