Add new icon tray-favourite.svg - SL #3868
Initially thought for Browse, to replace the tray button with a
simpler toogle button.  Adding a star to the tray graphic relates it
with bookmarks/favourites.  Then with Gary, humitos and Gonzalo we
found that this icon will be suitable to other activities that have
bookmarks, like Read and Wikipedia.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>