Don't set color for GtkLabel, do it in containers instead - SL #3974
An example of why this is wrong is #3974, in which removing:

SugarAlert GtkLabel { color: @white; }

makes the buttons label recover the right color, as specified in the
rule .buttons {...}

The only exception is in the notebook tabs: setting color there
doesn't paint the labels, so we have to add a rule for labels in tabs,
and then add another rule to make sure labels inside buttons inside
tabs get the right black color.  This patch also adds a comment in the
CSS explaining that.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>