Style activities in home view for tap - click and hover feedback
For hover, change the previous outline to a background fill.  For tap
- click, add a grey background fill.  This change is consistent with
the style of buttons in toolbars, which have subtle hover feedback,
and more visible background change for tap - click feedback.

For choosing the grey tones, Gary used the 'strive for at least two
Munsell steps' guideline as per the HIG.  Assuming our current white
zoom view canvas backrounds, and 50% grey for activities without
Journal entries (the 50% grey is 3 too bright for a strict two Munsell
gap, but close enough). This gives good visibility for both hover,
press, and enough contrast during a press on a 50% grey.

Also, make the corners a bit more round.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>