Remove double line in menuitem separators
We used padding in menuitem separators to add space between the
separator line and the items.  The original Sugar theme had a space of 1
subcell above and below the line.

Padding worked fine in GTK+ packaged in F18.  At that time I first tried
margin, but didn't work.

Now with sugar-build in F20 we see a double line.  Painting the border
yellow we can see what is happening (see picture attached to this PR).

The proper fix should be changing padding to margin.  But turns out
margin is not working yet.  I'll report this to GTK+.  In the mean time,
I think loosing the space is a good compromise to prevent the ugly
double line.