Fix clipboard icon size -  Fixes #4751
The origi of this problem was apply the padding rule to all the objects
internal to a SugarRadioToolButton. SugarClipboardIcon inherit the rule
and is affected, because have a button and a label inside.
The label is invisible, but with the padding enlarge the button.

Manuel said asterisks should not be used in the css theme, but few are present,
this is one case one using the asterisk have unintended consequences.

Signed-off-by: Gonzalo Odiard <godiard@sugarlabs.org>