Gtk 3.14 regressions: use sugar assets for checkboxes and radio buttons
Changes in Gtk 3.14 [1] expect assets for checkboxes and radio buttons to be
included in the icon theme, not as the css background.  This commit uses the
`-gtk-icon-source` property to tell gtk to use our pre-exsisting assets.

This also removes the needs for hacks such as adding a border in css, namely
commit 27fac30cb028a7461f40da6765db13c017ad6f13.

How to test:

1.  Open "gtk3-widget-factory" or the network control panel
2.  Notice how the checkboxes have the simple sugar tick, not the complex
    GNOME tick.

[1] https://mail.gnome.org/archives/gtk-devel-list/2014-May/msg00020.html