Fix radio button graphic in RadioToolButton background
On newer gtk+ versions, the background property of the radio
selector is taken literally.  This leads to any radio/check
things (such as RadioToolButtons in the homeview) gaining
the radio button graphic as a tiled background.

To maintain compatibility with Gtk+ 3.8 (OLPC OS 13), the
background is only removed when compiled against Gtk 3.19+
(which is where the bug occurs).